@@ -4,7 +4,8 @@
   "description": "an angular.io component collection for leaflet",
   "main": "index.js",
   "scripts": {
-    "start": "tsc",
+    "clean": "rimraf dist",
+    "start": "npm run clean && tsc",
     "tsc": "tsc",
     "tsc:w": "tsc -w",
     "pretest": "node node_modules/alpagegen/index.js",
@@ -31,22 +32,23 @@
   "peerDependencies": {
     "@angular/core": ">=2.0.0",
     "@angular/common": ">=2.0.0",
-    "@angular/http": ">=2.0.0"
+    "@angular/http": ">=2.0.0",
+    "rxjs": ">=5.0.0"
   },
   "homepage": "https://github.com/elasticrash/ngx.leaflet.component#readme",
   "devDependencies": {
-    "@angular/common": "4.4.0-RC.0",
-    "@angular/compiler": "4.4.0-RC.0",
-    "@angular/core": "4.4.0-RC.0",
-    "@angular/http": "4.4.0-RC.0",
-    "@angular/platform-browser": "4.4.0-RC.0",
-    "@angular/platform-browser-dynamic": "4.4.0-RC.0",
+    "@angular/common": "5.0.0",
+    "@angular/compiler": "5.0.0",
+    "@angular/core": "5.0.0",
+    "@angular/http": "5.0.0",
+    "@angular/platform-browser": "5.0.0",
+    "@angular/platform-browser-dynamic": "5.0.0",
     "@types/geojson": "^1.0.2",
     "@types/leaflet": "^1.2.0",
     "leaflet": "^1.2.0",
     "reflect-metadata": "0.1.10",
-    "rxjs": "^5.4.3",
-    "zone.js": "^0.8.17",
+    "rxjs": "5.5.2",
+    "zone.js": "^0.8.18",
     "@compodoc/compodoc": "^1.0.0-beta.1",
     "@types/jasmine": "^2.5.53",
     "@types/node": "^8.0.28",
